+++ lib/generators/templates/factories/factory_bot_rspec.rb
@@ -1,0 +1,5 @@
+FactoryBot.use_parent_strategy = true
+
+RSpec.configure do |config|
+  config.include FactoryBot::Syntax::Methods
+end